pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hariadiarief <adi.arief77@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.0"
Flask-HTTPAuth = "^4.8.0"
Werkzeug = "^3.0.1"
PyJWT = "^2.8.0"
Flask-SQLAlchemy = "^3.1.1"
psycopg2 = "^2.9.9"
Flask-Migrate = "^4.0.5"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
